GraphQL app (#2)

+++ db/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, Column, DateTime, ForeignKey, Integer, String
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 
@@ -8,7 +8,7 @@ from db.database import Base
 class Driver(Base):
     __tablename__ = "drivers"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     name = Column(String, nullable=False)
     surname = Column(String, nullable=False)
     age = Column(Integer, nullable=False)
@@ -21,7 +21,7 @@ class Driver(Base):
 class Car(Base):
     __tablename__ = "cars"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     brand = Column(String, nullable=False)
     model = Column(String, nullable=False)
     year_of_production = Column(Integer, nullable=False)
@@ -37,7 +37,7 @@ class Car(Base):
 class Ticket(Base):
     __tablename__ = "tickets"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     driver_id = Column(Integer, ForeignKey("drivers.id"), nullable=False)
     car_id = Column(Integer, ForeignKey("cars.id"))
     fine = Column(Integer)
